finally fix wall tiling, use a WFC-like algo
